Created MockObject for MoodleManager. Added to implementation of it: RealMoodleManager for real API calls MoodleManagerStub for testing and Parallel Development purposes
Clean a bit the code and added comments too

+++ app/src/main/java/de/tum/in/tumcampus/models/MoodleUserCourses.java
@@ -2,11 +2,7 @@ package de.tum.in.tumcampus.models;
 
 import android.text.Html;
 
-import org.json.JSONException;
 import org.json.JSONObject;
-
-import java.net.MalformedURLException;
-import java.net.URL;
 
 /**
  * Created by enricogiga on 05/06/2015.
@@ -94,7 +90,7 @@ public class MoodleUserCourses extends  MoodleObject{
             this.showgrades = jsonObject.optBoolean("showgrades");
             this.summaryformat = jsonObject.optDouble("summaryformat");
             this.summary=jsonObject.optString("summary");
-            if (summaryformat == 1){
+            if (summaryformat.doubleValue() == 1){
                 this.summary=stripHtml(summary);
             }
             this.visible = jsonObject.optDouble("visible");
